requirements.txt
gradio==6.6.0
gradio_client==2.1.0
pypdf
python-docx
sentence-transformers
chromadb
numpy
